requirements.txt
appnope==0.1.4
arcade==2.6.17
asttokens==2.4.1
attrs==24.2.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
comm==0.2.2
Cython==3.0.11
debugpy==1.8.7
decorator==5.1.1
executing==2.1.0
idna==3.10
ipykernel==6.29.5
ipython==8.29.0
jedi==0.19.1
jupyter_client==8.6.3
jupyter_core==5.7.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
packaging==24.1
parso==0.8.4
pexpect==4.9.0
Pillow==9.3.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pyglet==2.0.dev23
Pygments==2.18.0
pymunk==6.4.0

python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytiled-parser==2.2.0
pyzmq==26.2.0
requests==2.32.3
six==1.16.0
stack-data==0.6.3
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==2.2.3
wcwidth==0.2.13
